+++ changelog.md
@@ -4,6 +4,7 @@
 
 * remove `config` parameter from `processEndpoint` function
 * remove roles from authorization in ApiEndpoint base class
+* remove unneeded 'use' Firebase from Api_v2
 
 
 ### v1.3.0